requirements.txt
aiofiles==24.1.0
aiosmtpd==1.4.6
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
atpublic==5.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.4.1
fastapi==0.114.2
fastapi-cli==0.0.5
fastapi-utils==0.7.0
fastapi_amis_admin==0.7.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.9
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psutil==5.9.8
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.19.0
SQLAlchemy==2.0.34
sqlalchemy_database==0.1.2
sqlmodel==0.0.22
starlette==0.38.5
tomli==2.0.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
WTForms==3.1.2
